refactor: fix folder service to services
